Develop/home page

+++ html/views/delete.php
@@ -3,7 +3,7 @@ use Classes\Post\GetPost;
 use Classes\Post\DeletePost;
 
 print_r($_POST);
-if (!empty($_POST) && $_POST['btn_submit'] == 'delete') {
+if (isset($_POST['btn_submit']) && !empty($_POST) && $_POST['btn_submit'] == 'delete') {
     $delete_id = fun_h($_POST['delete_post_id']);
     $delete_post = new DeletePost($delete_id);
     $flag = $delete_post->deletePost();
